git-svn-id: file:///Users/ckoch/repos/assembla.com/boo@1285 e0e46c49-be69-4f5a-ad62-21024a331aea
Former-commit-id: cf39ff4dc7e80b05315a6fdd3939d7b435fafa6b

+++ dbtoaster/applications/trading/alpha2/StockExchangeSimOpt/src/state/OrderBook.java
@@ -108,7 +108,7 @@ public class OrderBook {
         schemaKeys.add("trader");
         
         schema = new HashMap<String, String>();
-        schema.put("timestamp", "int");
+        schema.put("timestamp", "long");
         schema.put("price", "double");
         schema.put("volume", "int");
         schema.put("order_id", "int");
